requirements.txt
transformers==4.34.0
torch==2.1.0
pandas==2.1.1
scipy==1.11.3
python-dotenv==1.0.0
anthropic==0.6.0
langchain==0.0.352
boto3==1.29.3
openai==1.6.1
cryptography==41.0.7